adding tests and separating the interpolator class
